Update the CI to include asan and ubsan support (#80)
* Update the CI to include asan and ubsan support

* Run in parallel, fix clang args format

* Add ASAN to presets and make it work in the CI

* Skip CI on blocked PRs

* Create sources/playground.cpp

* Update CMakeLists.txt in tests

* split ECPPS into lib and cli

* move main.cpp to cli/

* update #includes in main.cpp

* Delete compiler/ecpps/src/main.cpp

* Update CMakeLists.txt in tests

* Update test_compiler.cpp

* Update main.cpp

* Update cmake-multi-platform.yml

* Update CMakeLists.txt

* Update CMakeLists.txt

* Update cmake-multi-platform.yml

* Update CMakeLists.txt

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update PseudoAssembly.cpp

* Update ABI.cpp

* Update CI ctest paths


* Update CI ctest paths


* CI ASan


* CI ASan


* CI ASan, reserve calls and basic linux support


* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update CMakePresets.json

* Update cmake-multi-platform.yml

* Update CMakePresets.json

* Update cmake-multi-platform.yml

* Update CMakePresets.json

* Update Machine.h

* Make MSVC acutally use CL


* Fix pointer comparison (still not fully fixed but oh well)


* Delete a weird cache directory


* Fix MSVC case sensitivity


* Setup MSVC env


* Invisible spaces in CI


* Fix CI


* Cache LLVM and add ASan as a separate job


* Uncache LLVM (too slow...)


* rename


* Use clang-cl


* Fix warnings


* Update CMakeLists.txt

* First stdlib header


* Remove -Wc++20-compat


* Use presets for ASan


* Add ASan to build properties


* initial linux support (just compiles, the linker is not implemented)


* include <filesystem>


* Add fuzzer support


* Remove fuzzer


---------

+++ compiler/ecpps/src/Machine/Machine.h
@@ -18,7 +18,7 @@ namespace ecpps::abi
           ARM64,
      };
 
-     enum struct SimdFeatures : std::uint_fast16_t
+     enum struct SimdFeatures : std::uint_fast32_t
      {
           None = 0,
